requirements.txt
# AI Models
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
sentencepiece>=0.1.99
diffusers>=0.21.0

# Image Processing
Pillow>=9.0.0

# GUI
customtkinter>=5.2.0

# Web Server
Flask>=2.3.0
Flask-CORS>=4.0.0

# Database (for backend)
psycopg2-binary>=2.9.0
bcrypt>=4.0.0

# HTTP Requests
requests>=2.31.0 